Fixed database migrations and renamed files
