+++ server/controllers/userController.js
@@ -3,7 +3,7 @@ import bcrypt from 'bcrypt'
 import { validateUser } from '../util/validations'
 import JRes from '../util/JResponse'
 import Helpers from '../util/Helpers'
-import Model from '../config/db'
+import Model from '../config/Database'
 
 export default class UserController {
   /**
