+++ server/routes/authRoutes.js
@@ -1,6 +1,6 @@
 'use strict'
 
-import Auth from '../controllers/authController'
+import Auth from '../controllers/AuthController'
 
 module.exports = (router) => {
   router.post('/auth/login', Auth.login)
